+++ ChangeLog.md
@@ -6,6 +6,7 @@ All notable changes in `sebastianbergmann/environment` are documented in this fi
 
 ### Added
 
+* Implemented `Runtime::getNameWithVersionAndCodeCoverageDriver()` method
 * Implemented [#34](https://github.com/sebastianbergmann/environment/pull/34): Support for PCOV extension
 
 ## [4.0.2] - 2019-01-28